Rename functions and moved them to `common.go`
