+++ pkg/iac/terraform/state/enumerator/azurerm.go
@@ -60,7 +60,7 @@ func (s *AzureRMEnumerator) Origin() string {
 func (s *AzureRMEnumerator) Enumerate() ([]string, error) {
 
 	// prefix should contain everything that does not have a glob pattern should be the glob matcher string
-	prefix, pattern := GlobS3(s.objectPath)
+	prefix, pattern := SplitPath(s.objectPath)
 	fullPattern := strings.Join([]string{prefix, pattern}, "/")
 	fullPattern = strings.Trim(fullPattern, "/")
