+++ pkg/iac/terraform/state/enumerator/gs.go
@@ -45,7 +45,7 @@ func (s *GSEnumerator) Enumerate() ([]string, error) {
 	bucketName := bucketPath[0]
 	// prefix should contains everything that does not have a glob pattern
 	// Pattern should be the glob matcher string
-	prefix, pattern := SplitS3AndGSPath(strings.Join(bucketPath[1:], "/"))
+	prefix, pattern := SplitPath(strings.Join(bucketPath[1:], "/"))
 
 	fullPattern := strings.Join([]string{prefix, pattern}, "/")
 	fullPattern = strings.Trim(fullPattern, "/")
